Add basic gm logic

+++ gm/src/main/kotlin/demo/gos/gm/Element.kt
@@ -1,3 +1,3 @@
 package demo.gos.gm
 
-data class Element(val id: String, val type: ElementType, val x: Int, val y: Int, val status: ElementStatus?)
+data class Element(val id: String, val type: ElementType, val x: Double, val y: Double, val status: ElementStatus?)
